[UNSTABLE] Cents instead of dollars
* !! The balance PreferencesDataStore is not migrated in this version and the file must be manually deleted after updating to prevent crashes: /data/data/com.example.transactions/files/datastore/settings.preferences_pb
* Modified the database to use cents (Int // INTEGER) for amount fields instead of dollars (Double // REAL) due to inaccuracies with floating point values, migration from version 3 to 4 of the Room DB was provided.
* Fixed a bug where the first transaction of a recurring doesn't charge your balance when the recurring is first created

+++ app/src/main/java/com/example/transactions/data/Recurring.kt
@@ -9,7 +9,7 @@ data class Recurring (
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,
 
-    val amount: Double,
+    val amount: Int,
 
     val title: String,
